Adds duration property to QuamMacro and PulseMacro
Adds a `duration` property to the `QuamMacro` base class,
allowing subclasses to define their duration. It defaults to `None`,
indicating no fixed duration.

Implements the `duration` property in `PulseMacro` to
return the pulse length in seconds.

The main rational for switching it to a property is that it enables the end user to compute programmatically the duration of a macro based on the qubit information on which the macro is attached to, that would not be known at declaration time.
The property returns None by default and can be overriden in subclasses as shown in the PulseMacro example.