Renames duration property to inferred_duration
Renames the `duration` property to `inferred_duration` to better reflect its purpose and usage.

This change clarifies that the duration is not a user-defined value, but rather a value that is derived or inferred from other properties.
It also adds a default value for duration at the `QuamMacro` level. This enables the duration to therefore remain saved in the Quam state